requirements.txt
Jinja2==3.1.4
matplotlib==3.9.2
pillow==11.0.0
python-dotenv==1.0.1
requests==2.31.0